LayoutTests/http/tests/security/contentSecurityPolicy/nonce-hiding-on-svg-script-expected.txt

This tests that nonce content attribute is removed from a SVG script element even if it was a pending resource.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testScript.getAttribute("nonce") is ""
PASS testScript.nonce is "abc"
PASS testScript.getAttribute("executed") is "true"
PASS successfullyParsed is true

TEST COMPLETE
